% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports_fn.R
\name{get_snotel_report}
\alias{get_snotel_report}
\title{Get SNOTEL Report}
\usage{
get_snotel_report(site)
}
\arguments{
\item{site}{A snotel site id.}
}
\value{
A rendered HTML report
}
\description{
Get SNOTEL Report
}
