% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports_fn.R
\name{get_usgs_report}
\alias{get_usgs_report}
\title{Get USGS Report}
\usage{
get_usgs_report(site)
}
\arguments{
\item{site}{A USGS site id.}
}
\value{
A rendered HTML report.
}
\description{
Get USGS Report
}
