\name{hydroApps-package}
\alias{hydroApps-package}
\alias{hydroApps}
\docType{package}
\title{Tools and models for hydrological applications}
\description{
Implementation of models for regional hydrological analysis in Italy.}
\details{
\tabular{ll}{
Package: \tab hydroApps\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2013-09-05\cr
License: \tab GPL (>=2)\cr
}
At the moment, the package provides a simple way to implement the SSEM (Spatially Smooth Estimation Method) model, also known as ARPIEM2012, for Regional Flood Frequency Analysis in the Piemonte region (Northwestern Italy). The model is based on the approach by \emph{Laio et al.} (2011) calibrated using the most up-to-date database of annual maxima provided in \emph{Barbero et al.} (2012). 
}
\author{
Daniele Ganora
Maintainer: Daniele Ganora <daniele.ganora@polito.it>
}
\references{
S.Barbero, M.Graziadei, M.Zaccagnino, A.Saladin, P.Claps, D.Ganora,
  F.Laio, e R.Radice. \emph{Catalogo delle portate massime annuali al colmo del bacino
  occidentale del Po}. ISBN 978-88-7479-112-5. Arpa Piemonte, 2012.
  \url{http://www.arpa.piemonte.it/pubblicazioni-2/pubblicazioni-anno-2012/catalogoflorabassa.pdf}
  
  F.Laio, D.Ganora, P.Claps, e G.Galeati. Spatially smooth regional estimation of the flood frequency curve
  (with uncertainty). \emph{Journal of Hydrology}, 408: 67--77, 2011.
 doi10.1016/j.jhydrol.2011.07.022.
 
}
\keyword{ package }
%\seealso{
% Other useful hydrological tools in the \code{\link{nsRFA}} package.
%}