% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_station.R
\docType{data}
\name{usgs_sites_df}
\alias{usgs_sites_df}
\title{USGS sites.}
\format{
A data frame with 8348 rows and 24 variables:
\describe{
  \item{status}{shows whether current or historic}
  ...
}
}
\usage{
usgs_sites_df
}
\description{
A dataset containing all the USGS sites (historic and current) from all 50 states.
}
\keyword{datasets}
